Clarifying of a NetCatKS core, now we providing a more clear way to define our custom API's. All public interfaces was moved from Dispatcher API to Component API. Changes inside netcatks too. Introduce of BaseAPI and BaseRootAPI - a short way of defining API's for DProtocol structures. All paths now are absolute.
