+++ NetCatKS/Components/api/interfaces/__init__.py
@@ -1,13 +1,15 @@
 __author__ = 'dimd'
 
-from NetCatKS.Components.api.interfaces.adapters import IDynamicAdapterFactory
-from NetCatKS.Components.api.interfaces.loaders import IBaseLoader
+from NetCatKS.Components.api.interfaces.adapters import IDynamicAdapterFactory, IXMLResourceAPI
+from NetCatKS.Components.api.interfaces.adapters import IJSONResourceAPI, IJSONResourceRootAPI, IRequestSubscriber
+from NetCatKS.Components.api.interfaces.loaders import IBaseLoader, IXMLResource
 from NetCatKS.Components.api.interfaces.registration.adapters import IRegisterAdapters
 from NetCatKS.Components.api.interfaces.registration.factories import IRegisterFactories
 from NetCatKS.Components.api.interfaces.registration.protocols import IRegisterProtocols
 from NetCatKS.Components.api.interfaces.registration.wamp import IWAMPResource, IRegisterWamp, IWAMPComponent
 from NetCatKS.Components.api.interfaces.storages import IStorageRegister
-from NetCatKS.Components.api.interfaces.loaders import IUserFactory, IUserStorage, IUserWampComponent, IUserGlobalSubscriber
+from NetCatKS.Components.api.interfaces.loaders import IUserFactory, IUserStorage, IUserWampComponent
+from NetCatKS.Components.api.interfaces.loaders import IUserGlobalSubscriber, IJSONResource
 
 __all__ = [
     'IDynamicAdapterFactory',
@@ -22,5 +24,11 @@ __all__ = [
     'IUserFactory',
     'IUserStorage',
     'IUserWampComponent',
-    'IUserGlobalSubscriber'
+    'IUserGlobalSubscriber',
+    'IJSONResource',
+    'IJSONResourceAPI',
+    'IJSONResourceRootAPI',
+    'IRequestSubscriber',
+    'IXMLResourceAPI',
+    'IXMLResource'
 ]
