+++ NetCatKS/Components/api/public/__init__.py
@@ -1,5 +1,5 @@
 __author__ = 'dimd'
 
-from ...common import *
-from ..implementers import *
-from ..interfaces import *
+from NetCatKS.Components.common import *
+from NetCatKS.Components.api.implementers import *
+from NetCatKS.Components.api.interfaces import *
