+++ NetCatKS/Config/api/implementers/__init__.py
@@ -72,5 +72,9 @@ class Config(object):
 
 
 __all__ = [
-    'Config'
+    'Config',
+    'TCP',
+    'WS',
+    'WAMP',
+    'WEB'
 ]
