+++ NetCatKS/Config/api/implementers/configuration/tcp/__init__.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 __author__ = 'dimd'
 
 from zope.interface import implementer
@@ -5,7 +7,7 @@ from zope.component import getGlobalSiteManager
 from zope.component.factory import Factory
 from zope.component.interfaces import IFactory
 
-from NetCatKS.DProtocol import BaseProtocolActions
+from NetCatKS.DProtocol.api.public.actions import BaseProtocolActions
 from NetCatKS.Config.api.interfaces import ITcp
