+++ NetCatKS/Dispatcher/api/public/__init__.py
@@ -1,23 +1,12 @@
 __author__ = 'dimd'
 
-from NetCatKS.Dispatcher.api.interfaces import IJSONResource, IXMLResource, IHTMLResource, IBaseResourceSubscriber
-from NetCatKS.Dispatcher.api.interfaces import IJSONResourceAPI, IJSONResourceSubscriber
-from NetCatKS.Dispatcher.api.interfaces import IXMLResource, IXMLResourceAPI, IXMLResourceSubscriber
 from NetCatKS.Dispatcher.api.interfaces.dispatcher import IDispatcher, IDispathcherResultHelper
 from NetCatKS.Dispatcher.api.implementers.dispatcher import Dispatcher
 from NetCatKS.Dispatcher.api.implementers.dispatcher import DispathcherResultHelper
 
 __all__ = [
-    'IJSONResource',
-    'IXMLResource',
-    'IHTMLResource',
     'IDispatcher',
     'Dispatcher',
-    'IJSONResourceAPI',
-    'IJSONResourceSubscriber',
-    'IXMLResourceAPI',
-    'IXMLResourceSubscriber',
-    'IBaseResourceSubscriber',
     'DispathcherResultHelper',
     'IDispathcherResultHelper'
 ]
