--- NetCatKS/Validators/api/implementers/validators/html/__init__.py
@@ -1,62 +1,0 @@
-__author__ = 'dimd'
-
-
-from ....interfaces.message import IMessage
-from ..default import BaseValidator
-
-from zope.component import adapts
-from zope.component import getGlobalSiteManager
-
-from lxml.html import document_fromstring
-
-
-class HTMLValidator(BaseValidator):
-    """
-
-    Class provide functionality do detect incoming data for html format. HTMLValidator inherit
-    from BaseValidator and it's self provide IValidator and adapter for IMessage
-
-    All sub validators must be registered as SubscriptionAdapter
-    """
-    adapts(IMessage)
-
-    def __init__(self, validate_msg):
-        super(HTMLValidator, self).__init__(validate_msg)
-
-    def validate(self):
-
-        """
-
-        Will trying to load data as html, then will check for head and body tags, if they exist
-        is_valid will be true, message_type = 'HTML' and message = self.validate_msg.message
-
-        :return: self
-
-        """
-        try:
-            __html = document_fromstring(self.validate_msg.message)
-
-        except Exception as e:
-
-            return self
-
-        else:
-
-            try:
-
-                head = __html.head
-                body = __html.body
-
-            except IndexError:
-                return self
-
-            else:
-
-                self.is_valid = True
-                self.message_type = 'HTML'
-                self.message = self.validate_msg.message
-
-                return self
-
-gsm = getGlobalSiteManager()
-gsm.registerSubscriptionAdapter(HTMLValidator)
